fix: error caused of moving components dir
