+++ src/pages/component/usage-type-btn.tsx
